+++ examples/ethernetip/multi_position_nonblock.py
@@ -1,0 +1,35 @@
+import time
+from contextlib import ExitStack
+import logging
+from edcon.edrive.com_ethernetip import ComEthernetip
+from edcon.edrive.motion_handler import MotionHandler
+from edcon.utils.logging import Logging
+
+# Enable loglevel info
+Logging()
+
+# Create a list of all Modbus targets
+edrives = [ComEthernetip(ip_address='192.168.0.2'),
+           ComEthernetip(ip_address='192.168.0.51')]
+
+with ExitStack() as stack:
+    mots = [stack.enter_context(MotionHandler(edrive)) for edrive in edrives]
+
+    for mot in mots:
+        mot.acknowledge_faults()
+        mot.enable_powerstage()
+        if not mot.referenced():
+            mot.referencing_task()
+
+    for mot in mots:
+        mot.position_task(10000000, 300000, nonblocking=True)
+    time.sleep(0.1)
+
+    while True:
+        target_positions_reached = [
+            mot.target_position_reached() for mot in mots]
+        logging.info(f"Target positions reached: {target_positions_reached}")
+        if all(target_positions_reached):
+            break
+
+        time.sleep(0.1)