+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## v0.10.4 - 18.07.23
 ### Fixed
 - [Patch] Enabled possibility to create multiple ComEthernetip instances.
 - [Patch] Increased timeout for enabling the powerstage from 1 to 5 seconds.